+++ test/transform/resource/after-delombok/DataPlain.java
@@ -6,10 +6,6 @@ class Data1 {
 		this.x = x;
 	}
 	@java.lang.SuppressWarnings("all")
-	private Data1() {
-		this.x = 0;
-	}
-	@java.lang.SuppressWarnings("all")
 	public int getX() {
 		return this.x;
 	}
@@ -62,10 +58,6 @@ class Data2 {
 		this.x = x;
 	}
 	@java.lang.SuppressWarnings("all")
-	private Data2() {
-		this.x = 0;
-	}
-	@java.lang.SuppressWarnings("all")
 	public int getX() {
 		return this.x;
 	}
@@ -116,10 +108,6 @@ final class Data3 {
 	@java.lang.SuppressWarnings("all")
 	public Data3(final int x) {
 		this.x = x;
-	}
-	@java.lang.SuppressWarnings("all")
-	private Data3() {
-		this.x = 0;
 	}
 	@java.lang.SuppressWarnings("all")
 	public int getX() {