+++ test/transform/resource/after-delombok/DataWithGetter.java
@@ -7,10 +7,6 @@ class DataWithGetter {
 		this.z = z;
 	}
 	@java.lang.SuppressWarnings("all")
-	private DataWithGetter() {
-		this.z = null;
-	}
-	@java.lang.SuppressWarnings("all")
 	public void setX(final int x) {
 		this.x = x;
 	}