+++ test/transform/resource/after-delombok/InnerClass.java
@@ -20,10 +20,6 @@ class C {
 			this.a = a;
 		}
 		@java.lang.SuppressWarnings("all")
-		private D() {
-			this.a = null;
-		}
-		@java.lang.SuppressWarnings("all")
 		public A getA() {
 			return this.a;
 		}