+++ test/transform/resource/after-ecj/DataWithGetter.java
@@ -45,10 +45,6 @@
     super();
     this.z = z;
   }
-  private @java.lang.SuppressWarnings("all") DataWithGetter() {
-    super();
-    this.z = null;
-  }
   public @java.lang.SuppressWarnings("all") int getX() {
     return this.x;
   }