+++ test/transform/resource/after-ecj/InnerClass.java
@@ -49,10 +49,6 @@ class C {
       super();
       this.a = a;
     }
-    private @java.lang.SuppressWarnings("all") D() {
-      super();
-      this.a = null;
-    }
   }
   C() {
     super();